Fix TypeScript Float32Array cast

+++ src/app/page.tsx
@@ -119,7 +119,7 @@ export default function Home() {
     if (!floats.length) return;
 
     const buffer = ctx.createBuffer(1, floats.length, RELAY_OUTPUT_SAMPLE_RATE);
-    buffer.copyToChannel(floats, 0, 0);
+    buffer.copyToChannel(floats as Float32Array, 0, 0);
     const source = ctx.createBufferSource();
     source.buffer = buffer;
     source.connect(ctx.destination);
